+++ librelane/steps/openroad.py
@@ -239,6 +239,23 @@ class OpenROADStep(TclStep):
             pdk=True,
         ),
         Variable(
+            "SIGNAL_WIRE_RC_LAYERS",
+            Optional[List[str]],
+            "Sets estimated signal wire RC values to the average of these layers'. If you provide more than two, the averages are grouped by preferred routing direction and you must provide at least one layer for each routing direction.",
+            pdk=True,
+            deprecated_names=[
+                ("WIRE_RC_LAYER", lambda x: [x]),
+                ("DATA_WIRE_RC_LAYER", lambda x: [x]),
+            ],
+        ),
+        Variable(
+            "CLOCK_WIRE_RC_LAYERS",
+            Optional[List[str]],
+            "Sets estimated clock wire RC values to the average of these layers'. If you provide more than two, the averages are grouped by preferred routing direction and you must provide at least one layer for each routing direction.",
+            pdk=True,
+            deprecated_names=[("CLOCK_WIRE_RC_LAYER", lambda x: [x])],
+        ),
+        Variable(
             "PDN_CONNECT_MACROS_TO_GRID",
             bool,
             "Enables the connection of macros to the top level power grid.",
@@ -976,6 +993,10 @@ class STAPostPNR(STAPrePNR):
             for lef in extra_lefs:
                 lefs.append("--input-lef")
                 lefs.append(lef)
+        if pad_lefs := self.config["PAD_LEFS"]:
+            for lef in pad_lefs:
+                lefs.append("--input-lef")
+                lefs.append(lef)
         metrics_path = os.path.join(corner_dir, "filter_unannotated_metrics.json")
         filter_unannotated_cmd = [
             self.get_openroad_path(),
@@ -1190,6 +1211,77 @@ def _validate_io_ppl_mode(
 
 
 @Step.factory.register()
+class PadRing(OpenROADStep):
+    """
+    Assembles a pad ring on a floor-planned ODB file using OpenROAD's built-in pad placer.
+    """
+
+    id = "OpenROAD.PadRing"
+    name = "Pad Ring Generation"
+
+    config_vars = OpenROADStep.config_vars + [
+        Variable(
+            "PDN_CONNECT_MACROS_TO_GRID",
+            bool,
+            "Enables the connection of macros to the top level power grid.",
+            default=True,
+            deprecated_names=["FP_PDN_ENABLE_MACROS_GRID"],
+        ),
+        Variable(
+            "PDN_MACRO_CONNECTIONS",
+            Optional[List[str]],
+            "Specifies explicit power connections of internal macros to the top level power grid, in the format: regex matching macro instance names, power domain vdd and ground net names, and macro vdd and ground pin names `<instance_name_rx> <vdd_net> <gnd_net> <vdd_pin> <gnd_pin>`.",
+            deprecated_names=[("FP_PDN_MACRO_HOOKS", pdn_macro_migrator)],
+        ),
+        Variable(
+            "PDN_ENABLE_GLOBAL_CONNECTIONS",
+            bool,
+            "Enables the creation of global connections in PDN generation.",
+            default=True,
+            deprecated_names=["FP_PDN_ENABLE_GLOBAL_CONNECTIONS"],
+        ),
+        Variable(
+            "PAD_CFG",
+            Optional[Path],
+            "A custom pad configuration file. If not provided, the default pad config will be used.",
+        ),
+        Variable(
+            "PAD_SOUTH",
+            Optional[List[str]],
+            "The pad instance names for the south pad row.",
+        ),
+        Variable(
+            "PAD_EAST",
+            Optional[List[str]],
+            "The pad instance names for the east pad row.",
+        ),
+        Variable(
+            "PAD_NORTH",
+            Optional[List[str]],
+            "The pad instance names for the north pad row.",
+        ),
+        Variable(
+            "PAD_WEST",
+            Optional[List[str]],
+            "The pad instance names for the west pad row.",
+        ),
+    ]
+
+    def get_script_path(self):
+        return os.path.join(get_script_dir(), "openroad", "pad.tcl")
+
+    def run(self, state_in: State, **kwargs) -> Tuple[ViewsUpdate, MetricsUpdate]:
+        kwargs, env = self.extract_env(kwargs)
+        if self.config["PAD_CFG"] is None:
+            env["PAD_CFG"] = os.path.join(
+                get_script_dir(), "openroad", "common", "pad_cfg.tcl"
+            )
+            info(f"'PAD_CFG' not explicitly set, setting it to {env['PAD_CFG']}…")
+
+        return super().run(state_in, env=env, **kwargs)
+
+
+@Step.factory.register()
 class IOPlacement(OpenROADStep):
     """
     Places I/O pins on a floor-planned ODB file using OpenROAD's built-in placer.
@@ -1206,6 +1298,12 @@ class IOPlacement(OpenROADStep):
         + io_layer_variables
         + [
             Variable(
+                "IO_PIN_CORNER_AVOIDANCE",
+                Optional[Decimal],
+                "The distance from each corner within which pin placement should be avoided.",
+                units="µm",
+            ),
+            Variable(
                 "IO_PIN_PLACEMENT_MODE",
                 PPLMode,
                 "Decides the mode of the random IO placement option.",
@@ -1216,10 +1314,16 @@ class IOPlacement(OpenROADStep):
             Variable(
                 "IO_PIN_MIN_DISTANCE",
                 Optional[Decimal],
-                "The minimum distance between two pins. If unspecified by a PDK, OpenROAD will use the length of two routing tracks.",
-                units="µm",
+                "The minimum distance between two pins. The unit is microns or routing tracks, depending on whether IO_PIN_MIN_DISTANCE_IN_TRACKS is set. If unspecified by a PDK, OpenROAD will use the length of two routing tracks.",
+                units="µm or routing tracks",
                 pdk=True,
                 deprecated_names=["FP_IO_MIN_DISTANCE"],
+            ),
+            Variable(
+                "IO_PIN_MIN_DISTANCE_IN_TRACKS",
+                Optional[bool],
+                "Setting this variable to true allows IO_PIN_MIN_DISTANCE to be set in number of tracks instead of microns.",
+                pdk=True,
             ),
             Variable(
                 "IO_PIN_ORDER_CFG",
@@ -1805,16 +1909,6 @@ class DetailedRouting(OpenROADStep):
                 deprecated_names=["ROUTING_CORES"],
             ),
             Variable(
-                "DRT_MIN_LAYER",
-                Optional[str],
-                "An optional override to the lowest layer used in detailed routing. For example, in sky130, you may want global routing to avoid li1, but let detailed routing use li1 if it has to.",
-            ),
-            Variable(
-                "DRT_MAX_LAYER",
-                Optional[str],
-                "An optional override to the highest layer used in detailed routing.",
-            ),
-            Variable(
                 "DRT_OPT_ITERS",
                 int,
                 "Specifies the maximum number of optimization iterations during Detailed Routing in TritonRoute.",
@@ -1839,6 +1933,18 @@ class DetailedRouting(OpenROADStep):
                 default=10,
                 units="%",
                 deprecated_names=["DRT_ANTENNA_MARGIN"],
+            ),
+            Variable(
+                "DRT_ANTENNA_REPAIR_JUMPER_ONLY",
+                bool,
+                "Only use jumpers to fix antenna violations. Cannot be used in conjunction with DRT_ANTENNA_REPAIR_DIODE_ONLY.",
+                default=False,
+            ),
+            Variable(
+                "DRT_ANTENNA_REPAIR_DIODE_ONLY",
+                bool,
+                "Only use antenna diodes to fix antenna violations. Cannot be used in conjunction with DRT_ANTENNA_REPAIR_JUMPER_ONLY.",
+                default=False,
             ),
             Variable(
                 "DRT_SAVE_DRC_REPORT_ITERS",