+++ test/config/test_variable.py
@@ -19,6 +19,9 @@ from pyfakefs.fake_filesystem_unittest import Patcher
 from typing import Dict, List, Literal, Optional, Tuple, Type, Union
 
 
+pytestmark = pytest.mark.all
+
+
 @pytest.fixture
 def _mock_fs():
     with Patcher() as patcher: